requirement.txt
apache-airflow==2.6.3
apache-airflow-providers-amazon==7.3.0
boto3==1.26.50
pandas==1.5.3
numpy==1.24.2
requests==2.28.2
psycopg2-binary==2.9.5
sqlalchemy==2.0.4
pyarrow==11.0.0
python-dotenv==1.0.0